+++ java/byps/src-ser-json/com/wilutions/byps/JSerializer_Array.java
@@ -7,14 +7,14 @@ public abstract class JSerializer_Array extends BSerializer {
 	}
 
 	@Override
-	public Object read(final Object obj1, final BInput bin1, final int version) throws BException {
+	public Object read(final Object obj1, final BInput bin1, final long version) throws BException {
 		final BInputJson bin = ((BInputJson)bin1);
 		Object obj = internalRead(bin);
 		return obj;
 	}	
 	
 	@Override
-	public void write(Object obj1, BOutput bout1, int version) throws BException {
+	public void write(Object obj1, BOutput bout1, long version) throws BException {
 		final BOutputJson bout = (BOutputJson)bout1;
 		final BBufferJson bbuf = (BBufferJson)bout.bbuf;
 		internalWrite(obj1, bout, bbuf);
