+++ java/byps/src/com/wilutions/byps/BProtocolS.java
@@ -7,7 +7,7 @@ public class BProtocolS extends BProtocol {
 	
 	public static final BBinaryModel BINARY_MODEL = BBinaryModel.MEDIUM;
 
-	public BProtocolS(BApiDescriptor apiDesc, int negotiatedVersion, ByteOrder negotiatedByteOrder) {
+	public BProtocolS(BApiDescriptor apiDesc, long negotiatedVersion, ByteOrder negotiatedByteOrder) {
 		super(apiDesc, negotiatedVersion, negotiatedByteOrder);
 	}
